pyproject.toml
[tool.poetry]
name = "redis-logs"
version = "0.1.0"
description = ""
authors = ["Marek Dlabacek <marek.dlabacek@wftech.cz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
redis = "5.2.0"
PyYAML = "6.0.2"
psycopg2-binary = "2.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# dev dependencies
[tool.poetry.dev-dependencies]
faker = "33.0.0"
redis = "5.2.0"